requirements.txt
xmltodict>=0.12.0
generateDS==2.29.24
lxml>=4.2.5
pytest>=3.10.0
requests>=2.20.0
faker>=0.9.2
pyhamcrest>=1.9.0
flask>=1.0.3
flask-restful>=0.3.7
SQLAlchemy>=1.3.4
mysql-connector-python>=8.0.16
pyOpenSSL>=19.1.0
uWSGI>=2.0.18
schemathesis>=3.0.7
google-cloud-bigquery